posts restrictions and more url paths fixing

+++ PAL - final project/pal/posts/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import redirect, render
 from django.http import HttpResponse
-
+from django.contrib.auth.decorators import login_required
 from posts.forms import PostForm
 from posts.models import Post, Comment
 
@@ -15,6 +15,7 @@ def index(request):
 
     return render(request, 'posts/home.html', context)
 
+
 def post(request, pk):
     post = Post.objects.get(id=pk)
     comments = Comment.objects.all().order_by('-created')
@@ -27,6 +28,7 @@ def post(request, pk):
     return render(request, 'posts/post.html', context)
 
 
+@login_required(login_url='user:sign-in')
 def create_post(request):
     form = PostForm()
 
@@ -36,7 +38,7 @@ def create_post(request):
         if form.is_valid():
             form.save()
 
-            return redirect('index')
+            return redirect('posts:index')
 
 
     context = {
@@ -46,9 +48,13 @@ def create_post(request):
     return render(request, 'posts/post_form.html', context)
 
 
+@login_required(login_url='user:sign-in')
 def update_post(request, pk):
     post = Post.objects.get(id=pk)
     form = PostForm(instance=post) # get the instance of the specific post - prefill the inputs in the form
+
+    if request.user != post.user:
+        return render(request, 'posts/not_allowed.html')
 
     if request.method == 'POST':
         form = PostForm(request.POST, instance=post) # if instance is not specified here again, it will create a new post instead of updating
@@ -56,7 +62,7 @@ def update_post(request, pk):
         if form.is_valid():
             form.save()
 
-            return redirect('index')
+            return redirect('posts:index')
             # return redirect(request.META.get('HTTP_REFERER')) # NOT WORKING!!!
 
     context = {
@@ -67,13 +73,17 @@ def update_post(request, pk):
     return render(request, 'posts/post_form.html', context)
 
 
+@login_required(login_url='user:sign-in')
 def delete_post(request, pk):
     post = Post.objects.get(id=pk)
+
+    if request.user != post.user:
+        return render(request, 'posts/not_allowed.html')
 
     if request.method == 'POST':
         post.delete()
 
-        return redirect('index')
+        return redirect('posts:index')
         # return redirect(request.META.get('HTTP_REFERER')) # NOT WORKING!!!
 
     context = {
